pyproject.toml
[tool.poetry]
name = "buko"
version = "0.1.0"
description = ""
authors = ["Roberto Rodríguez <roberto.rodriguez@intelygenz.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
pydantic = {extras = ["email"], version = "^1.9.0"}
motor = "^2.5.1"
python-dotenv = "^0.20.0"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptograpy"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
